Use the MacChanger struct's originalMac field to determine if the module has already been configured
This goal of this commit is to provide a simple way to deal with the
original mac resotation issue I was facing where the MacChanger struct
was being overwritten if called twice/more than once. This should lock
up users from doing that but provide them a way around it ( not locking
them into an original mac address when I could want to change it
outside my bettercap session) with a consistent-ish experience until I
figure something else out. Or someone else does.

It’s also late and I need to be studying for my windows administration
class midterm.

It’s supposed to be an Advanced Windows Administration class. But it’s
not.

Anyways. I’m sure this commit isn’t perfect. But. What code is? :joy: