fix gateway and arpspoof in android 7 nethunter
in android 7~10 based nethunter,  the default gateway can not be found in 'ip route', while the command 'getprop' is not available in nethunter terminal. Without defined the gw, arpspoof will treated the local device as gw hence it can not work.
To fix it, we need to manually added a 'default gw' line in 'ip route' results.